+++ HISTORY.rst
@@ -19,6 +19,7 @@ Release History
 - Fixed error in management callback on failed operation.
 - Default AMQP encoding of bytes is now a String type and a bytearray is a Binary type.
 - Added AMQP Array type and fixed Long type range validation.
+- Added `header` argument to Message and BatchMessage for setting a MessageHeader.
 
 
 0.1.0b5 (2018-04-27)